Split guess input logic for readability
