+++ bee-redux/src/features/guesses/components/guessInput/controls/GuessInputBackspace.tsx
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import { GuessInputContext } from "@/app/GuessInputProvider";
+import { GuessInputContext } from "@/features/guesses/providers/GuessInputProvider";
 
 export function GuessInputBackspace() {
   const { guessBackspace } = useContext(GuessInputContext);
